+++ infrastructure/db/question_solution_schema.py
@@ -1,5 +1,5 @@
 from infrastructure.db.database import Base
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Float, String, ForeignKey
 from sqlalchemy.dialects.postgresql import UUID
 import uuid
 
@@ -10,7 +10,7 @@ class QuestionSolution(Base):
     exam_solution_id = Column(UUID(as_uuid=True), ForeignKey("exam_solutions.id", ondelete="CASCADE"), nullable=False)
     question_template_id = Column(UUID(as_uuid=True), ForeignKey("question_templates.id", ondelete="CASCADE"), nullable=False)
     answer = Column(String(300), nullable=False)
-    score = Column(Integer)
+    score = Column(Float)
 
     def __init__(self, id, exam_solution_id, question_template_id, answer, score):
         self.id = id
